api/src/main/java/uk/gov/hmcts/fees2/register/util/SecurityUtils.java
package uk.gov.hmcts.fees2.register.util;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.stereotype.Service;
import uk.gov.hmcts.fees.register.api.repositories.IdamRepository;
import uk.gov.hmcts.reform.idam.client.models.UserInfo;

import java.util.Objects;

@Service
public class SecurityUtils {

    public static final String AUTHORISATION = "Authorization";

    private final IdamRepository idamRepository;

    @Autowired
    public SecurityUtils(final IdamRepository idamRepository) {
        this.idamRepository = idamRepository;
    }

    public UserInfo getUserInfo() {
        final String userToken = getUserToken();
        if (userToken == null) {
            return null;
        }
        return idamRepository.getUserInfo(userToken);
    }

    public String getUserToken() {
        Jwt jwt = (Jwt) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
        return jwt.getTokenValue();
    }

    public boolean isAuthenticated() {
        return Objects.nonNull(SecurityContextHolder.getContext().getAuthentication());
    }

}
